# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_igeargeek_sales-system-api_internal_models.RevenueSizeOption:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  github_com_igeargeek_sales-system-api_internal_models.SalesTarget:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      quarter:
        type: integer
      target_value:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
      year:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handlers.revenueSizeOptionForm:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
  internal_handlers.salesTargetForm:
    properties:
      quarter:
        type: integer
      target_value:
        type: integer
      year:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    CRM/sales-pipeline API. This spec currently covers the
    Prospect stage, Pipeline stage, and Dashboard summary routes
    as a scaffold — the annotation pattern on those handlers
    (see internal/handlers/prospect_stages.go, pipeline_stages.go,
    dashboard.go) extends to the rest of internal/routes/routes.go
    as they're annotated. Run `go run github.com/swaggo/swag/cmd/swag@v1.16.6
    init -g cmd/api/main.go -o docs --parseDependency --parseInternal` from the
    repo root to regenerate docs/ after editing annotations (the two flags are
    required to resolve gorm.DeletedAt across package boundaries).
  title: Sales System API
  version: "1.0"
paths:
...
  /admin/revenue-sizes:
    get:
      description: Returns every configured Company revenue bucket (active + inactive),
        ordered by name.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.RevenueSizeOption'
            type: array
      security:
      - BearerAuth: []
      summary: List revenue sizes
      tags:
      - admin/revenue-sizes
    post:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Revenue size fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.revenueSizeOptionForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.RevenueSizeOption'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a revenue size
      tags:
      - admin/revenue-sizes
  /admin/revenue-sizes/{id}:
    delete:
      description: 'Admin-only. Soft-delete (is_active: false).'
      parameters:
      - description: Revenue size ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a revenue size
      tags:
      - admin/revenue-sizes
    patch:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Revenue size ID
        in: path
        name: id
        required: true
        type: integer
      - description: Revenue size fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.revenueSizeOptionForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.RevenueSizeOption'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a revenue size
      tags:
      - admin/revenue-sizes
  /admin/sales-targets:
    get:
      description: Admin-only. Returns every configured per-quarter/per-year sales
        target, optionally filtered to one year, ordered oldest-to-newest.
      parameters:
      - description: Filter to a single year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.SalesTarget'
            type: array
      security:
      - BearerAuth: []
      summary: List sales targets
      tags:
      - admin/sales-targets
    post:
      consumes:
      - application/json
      description: Admin-only. One row per (year, quarter); creating a second row
        for the same period is rejected — PATCH the existing one instead. quarter
        must be 1-4, year must be a valid year, and target_value must be non-negative.
      parameters:
      - description: Sales target fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.salesTargetForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.SalesTarget'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a sales target
      tags:
      - admin/sales-targets
  /admin/sales-targets/{id}:
    delete:
      description: Admin-only. A hard delete (unlike PipelineStage's soft-delete)
        — nothing else references a SalesTarget row by ID, and removing one just reverts
        that period to the flat AppSettings.QuarterlySalesTarget/4 fallback.
      parameters:
      - description: Sales target ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a sales target
      tags:
      - admin/sales-targets
    patch:
      consumes:
      - application/json
      description: Admin-only. Only target_value is editable in practice (year/quarter
        identify the row); changing year/quarter to collide with another existing
        row is rejected the same way Create is.
      parameters:
      - description: Sales target ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sales target fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.salesTargetForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.SalesTarget'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a sales target
      tags:
      - admin/sales-targets
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT, e.g. "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"